habit есть но есть проблемы

+++ lib/features/personalization/screens/habit/screens/habit_home.dart
@@ -1,3 +1,4 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:iconsax/iconsax.dart';
@@ -51,8 +52,19 @@ class _HabitHomeScreenState extends State<HabitHomeScreen> {
   DateTime _focusedDay = DateTime.now();
   DateTime? _selectedDay;
 
+  String? get userId {
+    final User? user = FirebaseAuth.instance.currentUser;
+    return user?.uid;
+  }
+
   @override
   Widget build(BuildContext context) {
+    if (userId == null) {
+      return Scaffold(
+        body: Center(child: Text('Пожалуйста, войдите в аккаунт')),
+      );
+    }
+
     final String selectedWeekDay = _selectedDay != null
         ? _weekDayToString(_selectedDay!.weekday)
         : _weekDayToString(DateTime.now().weekday);
@@ -60,7 +72,8 @@ class _HabitHomeScreenState extends State<HabitHomeScreen> {
     return Scaffold(
       backgroundColor: backgroundColor,
       appBar: AppBar(
-        title: Text('Habit Tracker', style: TextStyle(fontWeight: FontWeight.bold)),
+        title: Text('Habit Tracker',
+            style: TextStyle(fontWeight: FontWeight.bold)),
         centerTitle: true,
         backgroundColor: primaryColor,
         elevation: 5,
@@ -88,7 +101,8 @@ class _HabitHomeScreenState extends State<HabitHomeScreen> {
       body: Column(
         children: [
           Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
+            padding:
+                const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
             child: Container(
               decoration: BoxDecoration(
                 borderRadius: BorderRadius.circular(16),
@@ -145,13 +159,17 @@ class _HabitHomeScreenState extends State<HabitHomeScreen> {
             ),
           ),
           Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
+            padding:
+                const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
                 Text(
                   'Привычки на $selectedWeekDay',
-                  style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold, color: primaryColor),
+                  style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                      color: primaryColor),
                 ),
                 IconButton(
                   icon: Icon(Icons.refresh, color: primaryColor),
@@ -165,74 +183,133 @@ class _HabitHomeScreenState extends State<HabitHomeScreen> {
           ),
           Expanded(
             child: StreamBuilder<QuerySnapshot>(
-              stream: FirebaseFirestore.instance.collection('habits').snapshots(),
+              stream: FirebaseFirestore.instance
+                  .collection('users')
+                  .doc(userId)
+                  .collection('habits')
+                  .snapshots(),
               builder: (context, snapshot) {
                 if (!snapshot.hasData) {
                   return Center(child: CircularProgressIndicator());
                 }
+
                 final habits = snapshot.data!.docs;
 
+                // Текущая дата
+                final DateTime today = DateTime.now();
+                final String todayString =
+                    '${today.year}-${today.month.toString().padLeft(2, '0')}-${today.day.toString().padLeft(2, '0')}';
+                final String todayWeekDay = _weekDayToString(today.weekday);
+
+                // Фильтруем привычки
                 final filteredHabits = habits.where((habit) {
-                  final repeatDays = List<String>.from(habit['repeat'] ?? []);
-                  return repeatDays.contains(selectedWeekDay);
+                  final List<dynamic> repeatDays = habit['repeat'] ?? [];
+                  final List<dynamic> completedDates = habit['datesCompleted'] ?? [];
+
+                  // Дни недели, когда задача должна повторяться
+                  final bool isRepeatToday = repeatDays.contains(todayWeekDay);
+
+                  // Задача не должна отображаться, если:
+                  // - она относится к прошедшим датам
+                  // - и не выполнена
+                  final bool isTaskPast =
+                      completedDates.every((date) {
+                        final DateTime completedDate = DateTime.parse(date);
+                        return completedDate.isBefore(today);
+                      }) &&
+                          !isRepeatToday;
+
+                  return isRepeatToday && !isTaskPast;
                 }).toList();
 
                 if (filteredHabits.isEmpty) {
                   return Center(
                     child: Text(
-                      'Нет привычек для выбранного дня.',
+                      'Нет задач для выбранного дня.',
                       style: TextStyle(fontSize: 16, color: Colors.grey),
                     ),
                   );
                 }
 
                 return ListView.builder(
-                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                  padding: const EdgeInsets.fromLTRB(16.0, 0.0, 16.0, 70.0),
                   itemCount: filteredHabits.length,
                   itemBuilder: (context, index) {
                     final habit = filteredHabits[index];
+                    final List<dynamic> completedDates = habit['datesCompleted'] ?? [];
+                    final bool isCompletedToday = completedDates.contains(todayString);
 
-                    // Получаем цвет и значок из Firestore
-                    final colorValue = habit['color'] ?? 0xFF000DFF; // Если цвет отсутствует, используем основной
-                    final Color color = Color(colorValue); // Преобразуем значение цвета в Color
-                    final iconIndex = habit['icon'] ?? -1; // Проверяем индекс значка
+                    // Получаем цвет и значок привычки
+                    final colorValue = habit['color'] ?? 0xFF000DFF;
+                    final Color color = Color(colorValue);
+                    final iconIndex = habit['icon'] ?? -1;
                     final IconData icon = (iconIndex >= 0 && iconIndex < _icons.length)
                         ? _icons[iconIndex]
-                        : Icons.help_outline; // Значок по умолчанию, если индекс некорректен
+                        : Icons.help_outline;
 
-                    return Card(
-                      color: color.withOpacity(0.8), // Цвет карточки с прозрачностью
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-                      elevation: 2,
+                    return AnimatedContainer(
+                      duration: Duration(milliseconds: 500),
+                      curve: Curves.easeInOut,
+                      margin: EdgeInsets.only(bottom: 16),
+                      decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(12),
+                        gradient: LinearGradient(
+                          colors: isCompletedToday
+                              ? [Colors.grey, Colors.grey]
+                              : [color.withOpacity(1), color.withOpacity(0.1)],
+                          begin: Alignment.topLeft,
+                          end: Alignment.bottomRight,
+                        ),
+                        boxShadow: [
+                          BoxShadow(
+                            color: Colors.black12,
+                            blurRadius: isCompletedToday ? 10 : 6,
+                            offset: Offset(2, 4),
+                          ),
+                        ],
+                      ),
                       child: ListTile(
                         contentPadding: EdgeInsets.all(15),
-                        leading: Icon(
-                            icon,
-                            color: Colors.white,
-                            size: 24,
-                        ),
+                        leading: Icon(icon, color: Colors.white, size: 32),
                         title: Text(
                           habit['title'],
-                          style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+                          style: TextStyle(
+                            color: isCompletedToday ? Colors.grey.shade300 : Colors.white,
+                            fontWeight: FontWeight.bold,
+                          ),
                         ),
-                        trailing: Icon(Icons.arrow_forward_ios, color: Colors.white, size: 16),
-                        onTap: () {
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => AnalyticsScreen(habitId: habit.id),
-                            ),
-                          );
-                        },
+                        trailing: IconButton(
+                          icon: Icon(
+                            isCompletedToday ? Iconsax.activity : Icons.check,
+                            color: isCompletedToday ? Colors.grey.shade300 : Colors.white,
+                          ),
+                          onPressed: () async {
+                            if (isCompletedToday) {
+                              // Удаляем сегодняшнюю дату из списка выполненных
+                              completedDates.remove(todayString);
+                            } else {
+                              // Добавляем сегодняшнюю дату в список выполненных
+                              completedDates.add(todayString);
+                            }
+
+                            // Обновляем данные в Firestore
+                            await FirebaseFirestore.instance
+                                .collection('users')
+                                .doc(userId)
+                                .collection('habits')
+                                .doc(habit.id)
+                                .update({'datesCompleted': completedDates});
+                          },
+                        ),
                       ),
                     );
                   },
                 );
-
-
               },
             ),
           ),
+
+
         ],
       ),
     );
@@ -257,8 +334,11 @@ class _HabitHomeScreenState extends State<HabitHomeScreen> {
       ),
     );
 
-    if (confirmation == true) {
-      final collection = FirebaseFirestore.instance.collection('habits');
+    if (confirmation == true && userId != null) {
+      final collection = FirebaseFirestore.instance
+          .collection('users')
+          .doc(userId)
+          .collection('habits');
       final snapshots = await collection.get();
       for (var doc in snapshots.docs) {
         await doc.reference.delete();
@@ -271,7 +351,34 @@ class _HabitHomeScreenState extends State<HabitHomeScreen> {
   }
 
   String _weekDayToString(int weekDay) {
-    const days = ['Понедельник', 'Вторник', 'Среда', 'Четверг', 'Пятница', 'Суббота', 'Воскресенье'];
+    const days = [
+      'Понедельник',
+      'Вторник',
+      'Среда',
+      'Четверг',
+      'Пятница',
+      'Суббота',
+      'Воскресенье',
+    ];
     return days[weekDay - 1];
+  }
+
+// Функция для определения следующей даты, соответствующей дню недели
+  DateTime _nextWeekDayOccurrence(String weekDay) {
+    const daysMap = {
+      'Понедельник': 1,
+      'Вторник': 2,
+      'Среда': 3,
+      'Четверг': 4,
+      'Пятница': 5,
+      'Суббота': 6,
+      'Воскресенье': 7,
+    };
+
+    final int targetWeekDay = daysMap[weekDay]!;
+    final DateTime today = DateTime.now();
+    final int difference = (targetWeekDay - today.weekday + 7) % 7;
+
+    return today.add(Duration(days: difference == 0 ? 0 : difference));
   }
 }
